1) Avoid signed/unsigned comparisons. 2) Add iwyu violation for stdio.h to iwyu_globals.cc.
Yes, (2) is ironic.

R=wan
DELTA=7  (3 added, 0 deleted, 4 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=1617

+++ iwyu_preprocessor.cc
@@ -108,7 +108,7 @@ namespace {
 // tokens than expected, warn if the first one doesn't start "//".
 bool MatchOneToken(const vector<string>& tokens,
                    const string& token,
-                   int num_expected_tokens,
+                   size_t num_expected_tokens,
                    SourceLocation loc) {
   if (tokens.size() < num_expected_tokens) {
     return false;
@@ -132,7 +132,7 @@ bool MatchOneToken(const vector<string>& tokens,
 bool MatchTwoTokens(const vector<string>& tokens,
                     const string& token1,
                     const string& token2,
-                    int num_expected_tokens,
+                    size_t num_expected_tokens,
                     SourceLocation loc) {
   if (tokens.size() < num_expected_tokens) {
     return false;
